rename DayTest.java to CalendarTest.java
